+++ frontend/lost/src/containers/DataSources/EditDSModal.jsx
@@ -13,7 +13,7 @@ import { faAws, faMicrosoft } from '@fortawesome/free-brands-svg-icons'
 import { faFile, faNetworkWired, faSave } from '@fortawesome/free-solid-svg-icons'
 import { useEffect, useState } from 'react'
 import * as fbAPI from '../../api/fb'
-import { useOwnUser } from '../../actions/user/user_api'
+import { useOwnUser } from '../../api/user'
 import BaseModal from '../../components/BaseModal'
 import LostFileBrowser from '../../components/FileBrowser/LostFileBrowser'
 import * as Notification from '../../components/Notification'
