+++ frontend/lost/src/containers/Instruction/Instruction.jsx
@@ -11,7 +11,7 @@ import EditInstruction from './EditInstruction'
 import ViewInstruction from './ViewInstruction'
 import BaseContainer from '../../components/BaseContainer'
 import * as Notification from '../../components/Notification'
-import { useOwnUser } from '../../actions/user/user_api'
+import { useOwnUser } from '../../api/user'
 import { faUserPlus, faPen, faTrash, faEye } from '@fortawesome/free-solid-svg-icons'
 import CoreDataTable from '../../components/CoreDataTable'
 import { createColumnHelper } from '@tanstack/react-table'
